requirements.txt
psycopg [binary,pool] == 3.2.1
fastapi == 0.110.1
python-multipart == 0.0.12
pyjwt[crypto] == 2.3.0
bcrypt == 4.2.0
uvicorn == 0.29.0
redis == 5.0.4
aiofiles == 24.1.0
